feat: prom metrics (#47)
This PR adds MVP prometheus metrics reporting. It adds a `/metrics` endpoint that renders metrics in prometheus exposition format.

Due to the KV format we have to read every nft and deal to collect metrics. Since the size of the data is small right now it doesn't take long. The result is cached by cloudflare for 10 minutes.

It gives us the ability to build graphs for monitoring and alerting in Grafana:

<img width="1552" alt="Screenshot 2021-03-24 at 15 38 20" src="https://user-images.githubusercontent.com/152863/112339499-a0d2b780-8cb7-11eb-9151-4d4fd8f93a24.png">