fix: duplicate NFTs (#107)
Uploading or pinning the same file twice will result in 2 entries in the NFTS and NFTS_IDX KV. This PR explicitly handles this case and returns the existing NFT if it exists. For the pinning service API if the pin name or meta is different to what is stored then it is updated.

I believe I inadvertently introduced this bug when I removed the code in `models/nfts.set` that used to merge ALL data with an existing NFT. Continuing to merging all data would have also been problematic because the `created` date could be updated making it impossible to generate a key into the NFTS_IDX from the NFT.

Instead we're explicitly merging the parts of the NFT where we want to allow mutations so no surprises.