refactor: move pin status and size cron job to GH actions (#142)
This PR moves the cron job that follows up on pin status and size to Github actions.

The cron job in cloudflare is running every minute, but is limited to 50 subrequests, so can only followup on 50 at a time (max) and likely <25 if many of the items have been pinned.

This PR moves that cron job to a github action that runs every hour (for a max duration of 60 minutes). Usually it shouldn't take too long but if we have a big influx of pins via the pinning service API then it can take longer.