feat: allow extensions to ERC-1155 metadata (#146)
This PR allows extensions to the ERC-1155 metadata format. The metadata passed must still comply to ERC-1155, but it allows extensions such as [the ones OpenSea use](https://docs.opensea.io/docs/metadata-standards#section-metadata-structure) (`animation_url`, external_url, `attributes` etc.) to also be included in the metadata.

Note that the HTTP API already supports this, the change here is to make the JS client more permissive.