Adds global styling to api doc section
This adds a CSS file (I couldn't find a better way to customize swagger
ui) to the api docs section so it uses the fonts, colors and some
general styling from the rest of the site. This doesn't include all
styling; for instance the double buttons require changing the code
structure of swagger-ui, which is possible, but a more complex
solution. It does bring it closer to looking like the rest of site.

To correctly set the global CSS so it overrides swagger's existing
CSS I'm using double classes for each rule: `.swagger-ui.swagger-ui`.
This ensures the styles will be overridden without !important
specificity. This could be slightly improved with sass in the project
because I could wrap all the rules in this once with it's nesting
power.

Fixes #362.