feat: use Cluster statusAll method (#1265)
This PR switches to using the new `statusAll` method on cluster which now allows us to get the status for a given set of CIDs.

Due to nginx restrictions on request length we currently have to ask for 120 at a time. However it is significantly better than making 120 separate requests.