chore: inherit stdio (#1862)
Makes the console a bit more noisy but it is super helpful for when the process hangs.

These were hanging for me - I added this and it turns out docker was just taking ages to download images. I didn't know that was happening so I didn't wait around for it to finish.