fix: dag structure check (#2036)
`iterateDag` returns false if the DAG is incomplete, but also if it was unable to decode a block.

If the DAG is not complete, it is not necessarily "partial" we just might have not been able to decode a block.

That function needs to return the DAG structure - `false` is ambiguous. i.e. `false` could be `Partial` or `Unknown`.

This PR renames `iterateDag` to `getDagStructure` which returns the structure as `Complete`/`Partial`/`Unknown`.