chore: re-enable skipped test & workaround miniflare formdata issue (#2159)
This closes #2088 by re-enabling the test we were skipping, now that miniflare preserves the full file path.

It also works around a bug I found while updating. Miniflare v2.7+ fails when you provide an instance of `FormData` from `@web-std/form-data` as the body of a `dispatchFetch` call. It also seems to fail when using the `FormData` built into node 18+. It likes `undici`'s `FormData` though :smile: 

Miniflare updated their `undici` dependency in the 2.7 release, so that seems likely to be the cause of the issue. It manifests as the body being sent with a content-type of `text/plain;charset=utf-8` instead of `multipart/form-data`, so the api route handler doesn't know what to do with it.

In `nfts-upload.spec.js` we were constructing a `FormData` directly, so the fix is just to import `FormData` from undici instead of using the platform version. The tests in `nfts-store.spec.js` use the `Token.encode` fn from the client. I didn't want to make changes to the client just to make API tests happy, so I added a little helper to convert the `Token.encode` output to an undici `FormData`.

BTW, this will likely affect any clean builds from `main`, since we currently depend on miniflare `^2.5.0`, so if you don't have a lockfile it will resolve to the latest 2.x release.