feat: nft-store now encodes it's own damn files
nft-store was leaning on cluster to encode each file speperately, to build up the metadata.json for the user.

it now encodes the files to unixFS blocks itself. It accumlates the blocks in a MemoryBlockStore, creates the metadata and then exports it as a CAR file.

it then hands off to the regular uploadCarWithStat flow, to benefit from R2+S3 backups.

Note: It looks like we've been using hte wrong dagSize for everything uploaded via this enpoint. Uploading a single CBOR block CAR to cluster and asking it for the dag size just gives you the total bytes of the blocks in the CAR... not the size of the reachable dag from that root. So this is now more accurate! huzzah
see: https://github.com/ipfs-cluster/ipfs-cluster/blob/f376cf5106deeeb903b58e7e2431fa63bdea6900/adder/adder.go#L303

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>