feat: NFT.storage changes (#2560)
Make all (most?) of the requested changes from
https://github.com/orgs/w3s-project/projects/1/views/1?pane=issue&itemId=57594940

Note that once this PR is deployed, new user signups will be disabled -
I did not put these behind a feature flag as there doesn't seem to be a
strong need for one.

A few caveats:

1) I put the same banner at the top of every page - since the menu is in
the global layout it's tricky to get different banners displaying on
different pages. This means that the requests to have different pages
link to different URLs are challenging to accommodate, so I've gone down
the easy path.
2) The error message for new users is a little janky - I followed the
letter of the request but it's not terribly pretty:

<img width="1254" alt="Screenshot 2024-04-04 at 12 28 33 PM"
src="https://github.com/nftstorage/nft.storage/assets/1113/db13e99e-49f4-4453-bb90-e7c1d95afb62">

We could make this prettier but it's nontrivial work, so I'm proposing
we ship with this and iterate but would be open to iterating before we
ship.