feat: add test for w3up deal fetch
Add a test for fetching filecoin deal info from w3up.

This is definitely not my best work - ideally we'd abstract this w3up mocking logic in a way that makes it less stateful, but I'm hesitant to do that when we will likely not be making many changes here in the future.